fix(agents): address review feedback on toolbox commands
