+++ cli/azd/extensions/azure.ai.agents/internal/cmd/toolbox_connection.go
@@ -4,9 +4,7 @@
 package cmd
 
 import (
-	"context"
 	"fmt"
-	"slices"
 	"strings"
 
 	"azureaiagent/internal/exterrors"
@@ -34,184 +32,8 @@ and retargets the toolbox default.`,
 	return cmd
 }
 
-// connectionAddFlags carries the verb-specific flags for `connection add`.
-type connectionAddFlags struct {
-	index string
-}
-
-// newToolboxConnectionAddCommand returns the `connection add` command.
-func newToolboxConnectionAddCommand(extCtx *azdext.ExtensionContext) *cobra.Command {
-	extCtx = ensureExtensionContext(extCtx)
-	flags := &connectionAddFlags{}
-
-	cmd := &cobra.Command{
-		Use:   "add <toolbox> <connection>",
-		Short: "Attach a project connection to a toolbox.",
-		Long: `Attach a project connection to a toolbox.
-
-The tool entry shape is inferred from the connection's ARM category:
-  RemoteTool       → mcp tool with server_label/server_url from the connection target
-  CognitiveSearch  → azure_ai_search tool (requires --index)
-Other categories are rejected.
-
-If the toolbox has a local pending record (from 'toolbox create'), v1 is
-published with this connection as the only tool. Otherwise the current default
-version is fetched, the tool entry is appended, a new version is POSTed, and
-the toolbox default is retargeted.`,
-		Args: cobra.ExactArgs(2),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			return runConnectionAdd(
-				cmd.Context(), args[0], args[1], *flags,
-				readToolboxFlags(cmd, extCtx),
-				defaultConnectionResolver{},
-			)
-		},
-	}
-
-	cmd.Flags().StringVar(
-		&flags.index, "index", "",
-		"Index name (required when the connection's category is CognitiveSearch).",
-	)
-	registerToolboxOutputFlag(cmd)
-	return cmd
-}
-
-func runConnectionAdd(
-	ctx context.Context, toolboxName, connName string,
-	verb connectionAddFlags, parent toolboxFlags,
-	resolver connectionResolver,
-) error {
-	if err := validateToolboxName(toolboxName); err != nil {
-		return err
-	}
-	if err := validateOutputFormat(parent.output); err != nil {
-		return err
-	}
-	if strings.TrimSpace(connName) == "" {
-		return exterrors.Validation(
-			exterrors.CodeInvalidPositionalArg,
-			"<connection> must not be empty",
-			"pass the short name of a project connection",
-		)
-	}
-
-	client, resolved, err := resolveToolboxAndClient(ctx, parent)
-	if err != nil {
-		return err
-	}
-	logResolvedEndpoint("toolbox connection add", resolved)
-
-	store, closer, err := newAzdPendingToolboxStore()
-	if err != nil {
-		return exterrors.Internal(exterrors.CodeAzdClientFailed,
-			fmt.Sprintf("failed to open the pending-toolbox store: %s", err))
-	}
-	defer closer()
-
-	return runConnectionAddWith(ctx, client, resolver, store, resolved.Endpoint,
-		toolboxName, connName, verb, parent)
-}
-
-// runConnectionAddWith is the testable core.
-func runConnectionAddWith(
-	ctx context.Context, client toolboxClient, resolver connectionResolver,
-	store pendingToolboxStore,
-	endpoint, toolboxName, connName string,
-	verb connectionAddFlags, parent toolboxFlags,
-) error {
-	conn, err := resolver.resolveConnection(ctx, endpoint, connName)
-	if err != nil {
-		return err
-	}
-
-	entry, err := buildToolEntry(conn, verb.index)
-	if err != nil {
-		return err
-	}
-
-	// Pending-promotion path: if a pending record exists, POST v1 directly.
-	// A store-read failure must not silently fall through to the live-toolbox
-	// branch (which would 404 and report CodeToolboxNotFound).
-	pending, err := store.Get(ctx, endpoint, toolboxName)
-	if err != nil {
-		return exterrors.Internal(
-			exterrors.CodePendingToolboxStoreFailed,
-			fmt.Sprintf("failed to read pending toolbox state: %s", err),
-		)
-	}
-	if pending != nil {
-		req := &azure.CreateToolboxVersionRequest{
-			Description: pending.Description,
-			Tools:       []map[string]any{entry},
-		}
-		created, err := client.CreateToolboxVersion(ctx, toolboxName, req)
-		if err != nil {
-			return exterrors.ServiceFromAzure(err, exterrors.OpCreateToolboxVersion)
-		}
-		if _, err := store.Clear(ctx, endpoint, toolboxName); err != nil {
-			return exterrors.Internal(
-				exterrors.CodePendingToolboxStoreFailed,
-				fmt.Sprintf("failed to clear pending toolbox record: %s", err),
-			)
-		}
-		return emitConnectionAddResult(toolboxName, created.Version, conn, parent.output, true)
-	}
-
-	// Existing-toolbox path: fetch default → append → POST → PATCH default_version.
-	tb, err := client.GetToolbox(ctx, toolboxName)
-	if err != nil {
-		if isAzureNotFound(err) {
-			return exterrors.Dependency(
-				exterrors.CodeToolboxNotFound,
-				fmt.Sprintf("toolbox %q not found", toolboxName),
-				fmt.Sprintf(
-					"run 'azd ai agent toolbox create %q' first, then re-run 'connection add'",
-					toolboxName,
-				),
-			)
-		}
-		return exterrors.ServiceFromAzure(err, exterrors.OpGetToolbox)
-	}
-
-	current, err := client.GetToolboxVersion(ctx, toolboxName, tb.DefaultVersion)
-	if err != nil {
-		return exterrors.ServiceFromAzure(err, exterrors.OpGetToolboxVersion)
-	}
-
-	if duplicateConnectionInTools(current.Tools, conn.ID) {
-		return exterrors.Validation(
-			exterrors.CodeDuplicateConnection,
-			fmt.Sprintf(
-				"connection %q (%s) is already attached to toolbox %q",
-				connName, conn.ID, toolboxName,
-			),
-			fmt.Sprintf("use 'connection list %q' to inspect current tools", toolboxName),
-		)
-	}
-
-	newTools := slices.Clone(current.Tools)
-	newTools = append(newTools, entry)
-
-	req := &azure.CreateToolboxVersionRequest{
-		Description: current.Description,
-		Metadata:    current.Metadata,
-		Tools:       newTools,
-	}
-	created, err := client.CreateToolboxVersion(ctx, toolboxName, req)
-	if err != nil {
-		return exterrors.ServiceFromAzure(err, exterrors.OpCreateToolboxVersion)
-	}
-
-	if _, err := client.SetDefaultVersion(ctx, toolboxName, created.Version); err != nil {
-		return exterrors.ServiceFromAzure(err, exterrors.OpSetDefaultVersion)
-	}
-
-	return emitConnectionAddResult(toolboxName, created.Version, conn, parent.output, false)
-}
-
 // buildToolEntry returns the tool-entry map appropriate for the connection's
-// category. Enforces the --index flag rules from § 5.6 and the `tool.name`
-// regex from § 4.2.
+// category. Enforces the --index flag rules and the `tool.name` regex.
 func buildToolEntry(conn *projectConnection, index string) (map[string]any, error) {
 	if err := validateToolName(conn.Name); err != nil {
 		return nil, err
@@ -276,10 +98,12 @@ func buildToolEntry(conn *projectConnection, index string) (map[string]any, erro
 		return nil, exterrors.Validation(
 			exterrors.CodeUnsupportedConnectionCategory,
 			fmt.Sprintf(
-				"connection %q has category %q; v1 supports RemoteTool and CognitiveSearch only",
+				"connection %q has category %q which is not supported as a toolbox tool today; "+
+					"v1 supports RemoteTool (MCP) and CognitiveSearch (Azure AI Search) only",
 				conn.Name, conn.Category,
 			),
-			"use a RemoteTool (MCP) or CognitiveSearch (Azure AI Search) connection",
+			"use a RemoteTool (MCP) or CognitiveSearch (Azure AI Search) connection, "+
+				"or file an issue requesting support for the connection category you need",
 		)
 	}
 }
@@ -298,31 +122,29 @@ func duplicateConnectionInTools(tools []map[string]any, connID string) bool {
 	return found
 }
 
-// emitConnectionAddResult prints the standard output for a successful add.
-func emitConnectionAddResult(
-	toolboxName, newVersion string, conn *projectConnection, output string, promoted bool,
-) error {
-	if output == "json" {
-		payload := map[string]any{
-			"toolbox":             toolboxName,
-			"version":             newVersion,
-			"connection":          conn.Name,
-			"connectionId":        conn.ID,
-			"category":            string(conn.Category),
-			"promotedFromPending": promoted,
+// filterOutConnection returns tools[] with every entry whose
+// project_connection_id matches connID stripped (top-level and nested forms).
+// `removed` reports whether at least one entry was filtered.
+func filterOutConnection(tools []map[string]any, connID string) (result []map[string]any, removed bool) {
+	for _, t := range tools {
+		if toolEntryReferences(t, func(id string) bool { return id == connID }) {
+			removed = true
+			continue
 		}
-		return emitJSON(payload)
+		result = append(result, t)
 	}
-	if promoted {
-		fmt.Printf(
-			"Published toolbox %s version %s with connection %s.\n",
-			toolboxName, newVersion, conn.Name,
-		)
-	} else {
-		fmt.Printf(
-			"Attached connection %s to toolbox %s (now at version %s).\n",
-			conn.Name, toolboxName, newVersion,
-		)
+	return result, removed
+}
+
+// shortConnectionName extracts the connection's short name from the trailing
+// segment of its ARM ID (e.g. ".../connections/my-mcp" → "my-mcp"). Falls back
+// to the full id when no slash is present.
+func shortConnectionName(id string) string {
+	if id == "" {
+		return ""
 	}
-	return nil
+	if i := strings.LastIndex(id, "/"); i >= 0 && i < len(id)-1 {
+		return id[i+1:]
+	}
+	return id
 }
